Begin implementing SaveVTK
Going through process of porting original save solution callback to use vtk. we are redefining the entire thing so that we can more easily export all of our solution variables and not just a few. Small change in PointCloudDomain to track any changes. Used in original save callback but unused for us