Implement IGR
Rough implementation of IGR. Requires using an iterative solver that accepts a function instead of a matrix as actually simplifying the LHS system costs too much and needs to be done every iteration. Currently using gmres from IterativeSolvers with a fixed number of 10 iterations. Several improvements possible as this is mvp